virtio: move testing utilities in separate module
