+++ src/devices/src/virtio/mod.rs
@@ -16,6 +16,7 @@ mod mmio;
 pub mod net;
 pub mod persist;
 mod queue;
+pub mod test_utils;
 pub mod vsock;
 
 pub use self::balloon::*;
